Fix issue with radio causing MCU to sleep
