+++ src/apps/ping-pong/NucleoF446/radio_phy.cpp
@@ -15,128 +15,129 @@ ProtoWriteBuffer writeLoraBuffer;
 LoRaMessage loraMessage;
 LoRaAckMessage loraAckMessage;
 
-// int8_t lastRssiValue = 0;
-// int8_t lastSnrValue = 0;
+int8_t lastRssiValue = 0;
+int8_t lastSnrValue = 0;
 
 /*!
  * \brief Radio interupts
  */
-// void OnTxDone(void);
-// void OnRxDone(uint8_t *payload, uint16_t size, int16_t rssi, int8_t snr);
-// void OnTxTimeout(void);
-// void OnRxTimeout(void);
-// void OnRxError(void);
+void OnTxDone(void);
+void OnRxDone(uint8_t *payload, uint16_t size, int16_t rssi, int8_t snr);
+void OnTxTimeout(void);
+void OnRxTimeout(void);
+void OnRxError(void);
 
 /*!
  * Radio events function pointer
  */
-// static RadioEvents_t RadioEvents;
+static RadioEvents_t RadioEvents;
 
-// void InitRadioPhy() {
-//     // Radio initialization
-//     RadioEvents.TxDone = OnTxDone;
-//     RadioEvents.RxDone = OnRxDone;
-//     RadioEvents.TxTimeout = OnTxTimeout;
-//     RadioEvents.RxTimeout = OnRxTimeout;
-//     RadioEvents.RxError = OnRxError;
+void InitRadioPhy() {
+    // Radio initialization
+    RadioEvents.TxDone = OnTxDone;
+    RadioEvents.RxDone = OnRxDone;
+    RadioEvents.TxTimeout = OnTxTimeout;
+    RadioEvents.RxTimeout = OnRxTimeout;
+    RadioEvents.RxError = OnRxError;
 
-//     Radio.Init(&RadioEvents);
-//     Radio.SetChannel(RF_FREQUENCY);
+    Radio.Init(&RadioEvents);
+    Radio.SetChannel(RF_FREQUENCY);
 
-// #if defined(USE_MODEM_LORA)
+#if defined(USE_MODEM_LORA)
 
-//     Radio.SetTxConfig(MODEM_LORA, TX_OUTPUT_POWER, 0, LORA_BANDWIDTH,
-//                       LORA_SPREADING_FACTOR, LORA_CODINGRATE,
-//                       LORA_PREAMBLE_LENGTH, LORA_FIX_LENGTH_PAYLOAD_ON,
-//                       true, 0, 0, LORA_IQ_INVERSION_ON, 3000);
+    Radio.SetTxConfig(MODEM_LORA, TX_OUTPUT_POWER, 0, LORA_BANDWIDTH,
+                      LORA_SPREADING_FACTOR, LORA_CODINGRATE,
+                      LORA_PREAMBLE_LENGTH, LORA_FIX_LENGTH_PAYLOAD_ON,
+                      true, 0, 0, LORA_IQ_INVERSION_ON, 3000);
 
-//     Radio.SetRxConfig(MODEM_LORA, LORA_BANDWIDTH, LORA_SPREADING_FACTOR,
-//                       LORA_CODINGRATE, 0, LORA_PREAMBLE_LENGTH,
-//                       LORA_SYMBOL_TIMEOUT, LORA_FIX_LENGTH_PAYLOAD_ON,
-//                       0, true, 0, 0, LORA_IQ_INVERSION_ON, true);
+    Radio.SetRxConfig(MODEM_LORA, LORA_BANDWIDTH, LORA_SPREADING_FACTOR,
+                      LORA_CODINGRATE, 0, LORA_PREAMBLE_LENGTH,
+                      LORA_SYMBOL_TIMEOUT, LORA_FIX_LENGTH_PAYLOAD_ON,
+                      0, true, 0, 0, LORA_IQ_INVERSION_ON, true);
 
-//     Radio.SetMaxPayloadLength(MODEM_LORA, BUFFER_SIZE);
+    Radio.SetMaxPayloadLength(MODEM_LORA, BUFFER_SIZE);
 
-// #elif defined(USE_MODEM_FSK)
+#elif defined(USE_MODEM_FSK)
 
-//     Radio.SetTxConfig(MODEM_FSK, TX_OUTPUT_POWER, FSK_FDEV, 0,
-//                       FSK_DATARATE, 0,
-//                       FSK_PREAMBLE_LENGTH, FSK_FIX_LENGTH_PAYLOAD_ON,
-//                       true, 0, 0, 0, 3000);
+    Radio.SetTxConfig(MODEM_FSK, TX_OUTPUT_POWER, FSK_FDEV, 0,
+                      FSK_DATARATE, 0,
+                      FSK_PREAMBLE_LENGTH, FSK_FIX_LENGTH_PAYLOAD_ON,
+                      true, 0, 0, 0, 3000);
 
-//     Radio.SetRxConfig(MODEM_FSK, FSK_BANDWIDTH, FSK_DATARATE,
-//                       0, FSK_AFC_BANDWIDTH, FSK_PREAMBLE_LENGTH,
-//                       0, FSK_FIX_LENGTH_PAYLOAD_ON, 0, true,
-//                       0, 0, false, true);
+    Radio.SetRxConfig(MODEM_FSK, FSK_BANDWIDTH, FSK_DATARATE,
+                      0, FSK_AFC_BANDWIDTH, FSK_PREAMBLE_LENGTH,
+                      0, FSK_FIX_LENGTH_PAYLOAD_ON, 0, true,
+                      0, 0, false, true);
 
-//     Radio.SetMaxPayloadLength(MODEM_FSK, BUFFER_SIZE);
+    Radio.SetMaxPayloadLength(MODEM_FSK, BUFFER_SIZE);
 
-// #else
-// #error "Please define a frequency band in the compiler options."
-// #endif
-// }
+#else
+#error "Please define a frequency band in the compiler options."
+#endif
+}
 
-// void OnTxDone(void) {
-//     ApplyConfigIfPending();
+void OnTxDone(void) {
+    ApplyConfigIfPending();
 
-//     // if (!isExecutingCommand) {
-//     //     Radio.Rx(RX_TIMEOUT_VALUE);
-//     // } else {
-//     Radio.Sleep();
-//     // }
-// }
+    // if (!isExecutingCommand) {
+    //     Radio.Rx(RX_TIMEOUT_VALUE);
+    // } else {
+    Radio.Sleep();
+    // }
+}
 
-// void OnTxTimeout(void) {
-//     ApplyConfigIfPending();
+void OnTxTimeout(void) {
+    ApplyConfigIfPending();
 
-//     // if (!isExecutingCommand) {
-//     //     Radio.Rx(RX_TIMEOUT_VALUE);
-//     // } else {
-//     Radio.Sleep();
-//     // }
-// }
+    // if (!isExecutingCommand) {
+    //     Radio.Rx(RX_TIMEOUT_VALUE);
+    // } else {
+    Radio.Sleep();
+    // }
+}
 
-// void OnRxDone(uint8_t *payload, uint16_t size, int16_t rssi, int8_t snr) {
-//     // Track RSSI and SNR
-//     lastRssiValue = rssi;
-//     lastSnrValue = snr;
+void OnRxDone(uint8_t *payload, uint16_t size, int16_t rssi, int8_t snr) {
+    // Track RSSI and SNR
+    lastRssiValue = rssi;
+    lastSnrValue = snr;
 
-//     for (uint16_t i = 0; i < size; i++) {
-//         readLoraBuffer.push(payload[i]);
-//     }
+    for (uint16_t i = 0; i < size; i++) {
+        readLoraBuffer.push(payload[i]);
+    }
 
-//     auto result = loraMessage.deserialize(readLoraBuffer);
-//     if (result != ::EmbeddedProto::Error::NO_ERRORS) {
-//         return;
-//     }
+    auto result = loraMessage.deserialize(readLoraBuffer);
+    if (result != ::EmbeddedProto::Error::NO_ERRORS) {
+        Radio.Rx(RX_TIMEOUT_VALUE);
+        return;
+    }
 
-//     if (loraMessage.get_command() == LoRaMessage::CommandType::Configuration) {
-//         if (loraMessage.has_spreadingFactorConfig()) {
-//             auto config = loraMessage.get_spreadingFactorConfig();
-//             UpdateRadioSpreadingFactor(config.get_spreadingFactor(), true);
-//         }
-//         if (loraMessage.has_sequenceRequestConfig()) {
-//             SetSequenceRequestConfig(loraMessage.get_sequenceRequestConfig());
+    if (loraMessage.get_command() == LoRaMessage::CommandType::Configuration) {
+        if (loraMessage.has_spreadingFactorConfig()) {
+            auto config = loraMessage.get_spreadingFactorConfig();
+            UpdateRadioSpreadingFactor(config.get_spreadingFactor(), true);
+        }
+        if (loraMessage.has_sequenceRequestConfig()) {
+            SetSequenceRequestConfig(loraMessage.get_sequenceRequestConfig());
 
-//             // TODO send ACK if success
-//         }
-//     }
+            // TODO send ACK if success
+        }
+    }
 
-//     // Ensure that the message is not re-used
-//     loraMessage.clear();
-//     Radio.Rx(RX_TIMEOUT_VALUE);
-// }
+    // Ensure that the message is not re-used
+    loraMessage.clear();
+    Radio.Rx(RX_TIMEOUT_VALUE);
+}
 
-// void OnRxTimeout(void) {
-//     Radio.Rx(RX_TIMEOUT_VALUE);
-// }
+void OnRxTimeout(void) {
+    Radio.Rx(RX_TIMEOUT_VALUE);
+}
 
-// void OnRxError(void) {
-//     // if (IsGateWay) {
-//     //      printf("[Main] error\n");
-//     // }
-//     Radio.Rx(RX_TIMEOUT_VALUE);
-// }
+void OnRxError(void) {
+    // if (IsGateWay) {
+    //      printf("[Main] error\n");
+    // }
+    Radio.Rx(RX_TIMEOUT_VALUE);
+}
 
 void TransmitProtoBuffer() {
     Radio.Send(writeLoraBuffer.get_data(), writeLoraBuffer.get_size());
