+++ src/apps/ping-pong/NucleoF446/radio_phy.h
@@ -1,4 +1,5 @@
 #include "stdint.h"
+#include "radio.h"
 
 #ifndef RADIO_PHY__H__
 #define RADIO_PHY__H__
